# Samat_Check/FBA/portal/fba/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "FBA POS API",
    "version": "1.0"
  },
  "host": "FBAappuv01.fortebank.com:8082",
  "paths": {
...
    "/product": {
      "get": {
        "description": "Получение списка товаров",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Product"
        ],
        "operationId": "ProductGET",
        "parameters": [
          {
            "type": "int",
            "description": "номер страницы пагинатора",
            "name": "page",
            "in": "query"
          },
          {
            "type": "int",
            "description": "количество записей на странице",
            "name": "size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "JSON массив товаров",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Product"
              }
            }
          },
          "400": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "put": {
        "description": "Изменение товара",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Product"
        ],
        "operationId": "ProductPUT",
        "parameters": [
          {
            "description": "JSON для изменения модели товара",
            "name": "bodyjson",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JSON ответ на изменение товара",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          },
          "400": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "post": {
        "description": "Создание товара",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Product"
        ],
        "operationId": "ProductPOST",
        "parameters": [
          {
            "description": "JSON для создания товара",
            "name": "bodyjson",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JSON созданная модель товара",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          },
          "400": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/product/{id}": {
      "get": {
        "description": "Получение товара по ID",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Product"
        ],
        "operationId": "ProductGETByID",
        "parameters": [
          {
            "type": "string \u003cuuid4\u003e",
            "description": "идентификатор товара",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "JSON модель товара",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          },
          "400": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "delete": {
        "description": "Удаление товара по ID",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Product"
        ],
        "operationId": "ProductDELETE",
        "parameters": [
          {
            "type": "string \u003cuuid4\u003e",
            "description": "идентификатор товара",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Успешное удаление",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
    "/products/count": {
      "get": {
        "description": "Получение количества товаров",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Product"
        ],
        "operationId": "ProductCount",
        "responses": {
          "200": {
            "description": "JSON с количеством товаров",
            "schema": {
              "$ref": "#/definitions/Count"
            }
          },
          "500": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Category": {
      "description": "Category категории товаров",
      "type": "object",
      "properties": {
        "ID": {
          "type": "string",
          "format": "uuid4"
        },
        "Name": {
          "type": "string"
        },
        "ParentID": {
          "type": "string",
          "format": "uuid4"
        },
        "Staff": {
          "$ref": "#/definitions/Staff"
        },
        "StaffID": {
          "type": "string",
          "format": "uuid4"
        }
      },
      "x-go-package": "fba/dbmodel"
    },
...
    "Count": {
      "description": "modelRowsCount структура для количества записей под swagger",
      "type": "object",
      "properties": {
        "count": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Count"
        }
      },
      "x-go-name": "modelRowsCount",
      "x-go-package": "fba/model"
    },
...
    "Error": {
      "description": "Error Ошибка",
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error"
        }
      },
      "x-go-package": "fba/model"
    },
...
    "Product": {
      "description": "Product продукты",
      "type": "object",
      "properties": {
        "Barcode": {
          "type": "string"
        },
        "Category": {
          "$ref": "#/definitions/Category"
        },
        "CategoryID": {
          "type": "string",
          "format": "uuid4"
        },
        "ID": {
          "type": "string",
          "format": "uuid4"
        },
        "Name": {
          "type": "string"
        },
        "Staff": {
          "$ref": "#/definitions/Staff"
        },
        "StaffID": {
          "type": "string",
          "format": "uuid4"
        }
      },
      "x-go-package": "fba/dbmodel"
    },
...
    "Role": {
      "description": "Role права пользователей(staff)",
      "type": "object",
      "properties": {
        "ID": {
          "type": "string",
          "format": "uuid4"
        },
        "Name": {
          "type": "string"
        }
      },
      "x-go-package": "fba/dbmodel"
    },
    "Staff": {
      "description": "Staff пользователь",
      "type": "object",
      "properties": {
        "ID": {
          "type": "string",
          "format": "uuid4"
        },
        "Name": {
          "type": "string"
        },
        "ParentID": {
          "type": "string",
          "format": "uuid4"
        },
        "Role": {
          "$ref": "#/definitions/Role"
        },
        "RoleID": {
          "type": "string",
          "format": "uuid4"
        }
      },
      "x-go-package": "fba/dbmodel"
    },
...
  },
  "security": [
    {
      "api_key": []
    }
  ]
}